server/requirements.txt
langchain==0.2.9
langchain_community==0.2.7
langchain_core==0.2.20
langchain-google-vertexai==1.0.6
langchain-huggingface==0.0.3
python-dotenv==1.0.1
Flask==3.0.3
Flask-Cors==3.0.10
google-generativeai==0.7.2
google-auth==2.32.0
langchain-google-genai==1.0.7